+++ screamer/__init__.py
@@ -1,8 +1,8 @@
 # screamer/__init__.py
 
-from .screamer_bindings import Lag
-from .generators import lag_generator
+from .screamer_bindings import Lag, Diff
+from .generators import lag_generator, diff_generator
 
 __version__ = "0.1.31"
 
-__all__ = ["Lag", "lag_generator"]
+__all__ = ["Lag", "lag_generator", "Diff", "diff_generator"]