+++ screamer/generators.py
@@ -1,15 +1,25 @@
 # screamer/generators.py
 
-from .screamer_bindings import Lag
+from .screamer_bindings import Lag, Diff
 
-def lag_generator(iterable, delay, initial=0.0):
+def lag_generator(iterable, delay):
     """
     Generator that yields lagged values from the input iterable.
 
     :param iterable: An iterable of numerical values.
     :param delay: Number of steps to lag.
-    :param initial: Initial value(s) to yield before lagged values are available.
     """
-    lag = Lag(delay, initial)
+    lag = Lag(delay)
     for value in iterable:
         yield lag(value)
+
+def diff_generator(iterable, delay):
+    """
+    Generator that yields changed values from the input iterable.
+
+    :param iterable: An iterable of numerical values.
+    :param delay: Number of steps to lag.
+    """
+    diff = Diff(delay)
+    for value in iterable:
+        yield diff(value)