+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project are documented in this file.
 [Unreleased] - yyyy-mm-dd
 -------------------------
 
+### Added
+* Moving average: Ma, Ewma
+* Differences: Diff
+* Returns: Lret, Sret
+
+### Changed
+* removed initial_value from the constructor, we (for now) return NaN values when we cant resolve indicators.
+
 Version v0.1.31 (2024-10-20)
 -------------------------
 
@@ -14,4 +22,3 @@ Version v0.1.31 (2024-10-20)
 
 ### Fixed
 
-